+++ frontend/src/features/authentication/index.js
@@ -1,3 +1,4 @@
 import GoogleLoginButton from './components/GoogleLoginButton'
+import GoogleLogoutButton from './components/GoogleLogoutButton'
 
-export { GoogleLoginButton }
+export { GoogleLoginButton, GoogleLogoutButton }
